fix: rename to avoid conflicts with new tests.common.*
